feat: Allow skipping actions in interactive mode

+++ src/actions/file_replace.ts
@@ -1,13 +1,17 @@
 import fs from "fs-extra";
 import path from "path";
-import { CliTest } from "../clitest";
+import { CliTest, ConfirmAction } from "../clitest";
 import { ActionComplete } from "./action";
 
 export async function fileReplace(dt: CliTest, action: ActionComplete) {
     const content = action.lines.join("\n") + "\n";
     dt.commands(`\nCWD: ${dt.cwd}`);
     dt.commands(`Replacing file: ${action.params.file} with:\n\n${content}`);
-    await dt.userConfirm("Continue?");
+    const confirm = await dt.userConfirm("Continue?");
+    if (confirm === ConfirmAction.skip) {
+        dt.info(`SKIPPING: Replace file ${action.params.file}`);
+        return;
+    }
 
     await fs.writeFile(path.join(dt.cwd, action.params.file), content);
 }
